Updated the JobQueue class.
Renamed it to WorkQueue. Changed it to house it’s own state object and
to pass it, along with a reference to the done method into each job in
the queue when they get called. This way the module doesn’t have to
pass in the state object each time it adds something to the queue, or
when it is done. Renamed add method to append and added an insert
method to place a job anywhere in the queue. This is useful for a job
that wants to add something to happen directly after itself.